project is complete

+++ Server/src/main/java/edu/sdccd/cisc191/template/StoredNumbersInt.java
@@ -1,8 +1,11 @@
 package edu.sdccd.cisc191.template;
 
+/**
+ * implements methods in StoredNumbers class
+ */
 public interface StoredNumbersInt {
     void clicksPerSecond(int clickCounter, int player, int attemptNum);
-    int averageCPS(int player);
+    double averageCPS(int player);
 }
